requirements.txt
requests
pyjwt
unittest2
pytest
pytest-html
